# backend/swagger-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Course Final Project - Auction API",
    "description": "Complete API documentation for the online auction platform. This API supports product bidding, user authentication, messaging, and auction management.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "http://localhost:3000"
    },
    "license": {
      "name": "AGPL-3.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "JWT token for authentication"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "description": "Unique user identifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "full_name": {
            "type": "string",
            "description": "Full name of the user"
          },
          "address": {
            "type": "string",
            "description": "User's address"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "description": "Date of birth"
          },
          "role": {
            "type": "string",
            "enum": ["ADMIN", "SELLER", "BIDDER"],
            "description": "User role"
          },
          "is_verified": {
            "type": "boolean",
            "description": "Email verification status"
          },
          "positive_rating": {
            "type": "integer",
            "description": "Count of positive ratings"
          },
          "negative_rating": {
            "type": "integer",
            "description": "Count of negative ratings"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "integer",
            "description": "Unique product identifier"
          },
          "seller_id": {
            "type": "integer",
            "description": "ID of the seller"
          },
          "category_id": {
            "type": "integer",
            "description": "Category ID"
          },
          "winner_id": {
            "type": "integer",
            "description": "ID of the auction winner (null if ongoing)"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "price_start": {
            "type": "number",
            "format": "decimal",
            "description": "Starting bid price"
          },
          "price_step": {
            "type": "number",
            "format": "decimal",
            "description": "Minimum bid increment"
          },
          "price_buy_now": {
            "type": "number",
            "format": "decimal",
            "description": "Buy now price (optional)"
          },
          "price_current": {
            "type": "number",
            "format": "decimal",
            "description": "Current highest bid"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "description": "Auction start date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "description": "Auction end date"
          },
          "is_auto_extend": {
            "type": "boolean",
            "description": "Auto-extend auction if bid placed near end"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "ENDED", "SOLD"],
            "description": "Product auction status"
          }
        }
      },
      "ProductDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Product"
          },
          {
            "type": "object",
            "properties": {
              "description": {
                "type": "string",
                "description": "Detailed product description"
              },
              "images": {
                "type": "array",
                "description": "Product images",
                "items": {
                  "type": "object",
                  "properties": {
                    "image_id": {
                      "type": "integer"
                    },
                    "image_url": {
                      "type": "string"
                    },
                    "is_primary": {
                      "type": "boolean"
                    }
                  }
                }
              },
              "seller": {
                "$ref": "#/components/schemas/User"
              },
              "category": {
                "type": "object",
                "properties": {
                  "category_id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              },
              "current_bids": {
                "type": "array",
                "description": "List of bids placed on this product",
                "items": {
                  "$ref": "#/components/schemas/Bid"
                }
              }
            }
          }
        ]
      },
...
      "Bid": {
        "type": "object",
        "properties": {
          "bid_id": {
            "type": "integer",
            "description": "Unique bid identifier"
          },
          "product_id": {
            "type": "integer",
            "description": "Product being bid on"
          },
          "bidder_id": {
            "type": "integer",
            "description": "User placing the bid"
          },
          "amount": {
            "type": "number",
            "format": "decimal",
            "description": "Bid amount"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when bid was placed"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/products/": {
      "get": {
        "summary": "Get All Active Products",
        "operationId": "getAllProducts",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of products per page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of products to skip",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search query for product name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create New Product",
        "operationId": "createProduct",
        "tags": ["Products"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 5,
                    "example": "Vintage Camera"
                  },
                  "category_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "price_start": {
                    "type": "number",
                    "format": "decimal",
                    "example": 10.00
                  },
                  "price_step": {
                    "type": "number",
                    "format": "decimal",
                    "example": 1.00
                  },
                  "price_buy_now": {
                    "type": "number",
                    "format": "decimal",
                    "example": 50.00
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-01-01T00:00:00Z"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-02-01T00:00:00Z"
                  },
                  "is_auto_extend": {
                    "type": "boolean",
                    "example": true
                  },
                  "images": {
                    "type": "array",
                    "maxItems": 4,
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Product images (max 4)"
                  }
                },
                "required": ["name", "category_id", "price_start", "price_step", "end_date", "images"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Get Product Details",
        "operationId": "getProductDetail",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProductDetail"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Product",
        "operationId": "updateProduct",
        "tags": ["Products"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 5
                  },
                  "category_id": {
                    "type": "integer"
                  },
                  "price_start": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "price_step": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "price_buy_now": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "is_auto_extend": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Product",
        "operationId": "deleteProduct",
        "tags": ["Products"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comments/{id}": {
      "delete": {
        "summary": "Delete Comment",
        "operationId": "deleteComment",
        "tags": ["Comments"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Comment ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}